HIVE-27118: move test case to TestRemoteHiveMetaStoreKerberos
